refactor(cli): modularize command implementation
